fix: don't call cabi_post_* for primitives
Primitives don't allocate memory, so we don't need to call the cleanup
function. This was breaking instances where the return type was a
u32, since they would try to reference a memory location that wasn't
allocated.

The `needs_cleanup` method is somewhat incomplete because it doesn't
handle user-defined types yet, but that's inline with the rest of the
codebase for now.

For transparency, this was done with some help from Claude, so I did
actually generate a bit of a test harness, but don't want to bombard
the repo with too much opinionated code!