feat: add 'Code generated by' header to generated files
The [docs for the Go command][docs] say that generated files
should have a header that matches the regex:

    ^// Code generated by .* DO NOT EDIT\.$

This commit adds a header to the generated files that matches
that regex. It includes the command that generated the file
in the header along with the arguments, but perhaps we should
just use the command name and not the full args?

[docs]: https://pkg.go.dev/cmd/go#hdr-Generate_Go_files_by_processing_source