feat: support stable Rust
Prior to this commit, this project required nightly Rust.
This was because it used genco 0.17, which didn't generate
code with the correct indentation. However, since Rust 1.88 and
genco 0.18, this can be done on stable Rust.

This commit switches the rust-toolchain.toml to a stable toolchain,
removes the build script, and updates the README to use stable instead.