refactor: move various Go syntax related structs to library; add tests
This commit moves the various Go syntax related structs to a `go` module
in a library, adds unit tests for them, and refactors the `main.rs` file
to use the library. It doesn't yet do anything with the 'Func' struct
or a couple of others, but it's a step in the direction of a more
succinct `main.rs` file.