refactor: move Bindings struct to lib; begin refactoring codegen
This starts the refactoring of codegen parts into the library, where it
can be tested and more easily changed.

It moves the `Bindings` struct into the library, and adds a new `Wasm`
struct responsible for generating code to embed or inline the
WebAssembly file.

As part of this, I needed `GoIdentifier` to be an owned type, so I
changed it to hold strings instead of references. This doesn't really
affect performance because we were always passing references to the
output of `format!()` which would allocate a new string anyway.