refactor: only expect one world
Rather than iterating over the worlds and incurring another level of
indentation, use 'find' to get the matching world and panic if we
can't find it. We probably don't want to continue if there was no
matching world, so this should be fine.

I've extracted this out into its own commit and PR so that the
unindentation doesn't cause massive diffs later.