refactor: refactor import and factory generation
This commit refactors the import and factory generation code
to a more modular approach. There are now two steps:

1. The import analysis step, which analyzes the imports in the
   world and generates a list of interfaces and standalone types
   in an intermediate representation.
2. The import code generation step, which takes the intermediate
   representation plus the WIT functions, and generates the Go
   code for the imports.

The goal is to make the code more modular and easier to maintain
and test.

In doing so, some of the generated code is ever so slightly
different from the original, but I've tried to keep the changes
to an absolute minimum. Pretty much all that's changed is the order
in which the `Close` method is generated, I hope.