feat: partial support for resources
This adds support for resources to gravity. It's a pretty huge change
and I'm not sure it's ready for review yet (or maybe ever), but it's a
decent PoC. Much of the code was written by Claude with a huge amount of
guidance, so it will probably need some cleanup.

The main limitations are a lack of support for 'owned' resources as
parameters because it involves a lot of complexity re. tracking
ownership of resources. Even so this is quite useful for many things.

Relates to https://github.com/arcjet/gravity/issues/5.