refactor: move Bindings struct to lib & begin refactoring codegen (#107)
Start the refactoring of codegen parts into the library, where it
can be tested and more easily changed.

It moves the `Bindings` struct into the library, and adds a new `Wasm`
struct responsible for generating code to embed or inline the
WebAssembly file.

As part of this, `GoIdentifier` has become an owned type, so it was
changed to hold strings instead of references. This doesn't really
affect performance since references were always passed to
`format!()` which would allocate a new string anyway.

---------

Co-authored-by: blaine-arcjet <146491715+blaine-arcjet@users.noreply.github.com>