feat: handle both Wasm32 and Wasm64 architectures
This commit adds support for both Wasm32 and Wasm64 architectures.
This is done by adding a new field to the generated factory and instance
which holds the configured architecture, and using the option pattern
to let the user configure the factory, as well as providing a
'FactoryWithArchitecture' option which can be used to set the
architecture.

The architecture defaults to Wasm32, since this is the predominant
architecture for most of the WebAssembly ecosystem.

We use the `ArchitectureSize::format_term` and `SizeAlign::format`
methods provided by `witparser` to get a string representing the
expression used to calculate the pointer size and alignment,
using the architecture on the factory (for imports) and on the
instance (for exports).

See https://github.com/arcjet/gravity/issues/58.