feat: add more support for customizing factory
This adds some more Options that can be handled by the factory, such as:

- providing custom Wasm bytes, rather than using the embedded ones
- instantiating the WASI snapshot module in the runtime, required for
  WASI
- providing a custom module config, e.g. to enable stdin/stdout when
  using WASI, or to customise the 'Start' function used by Wazero

It also adds a new example which uses a guest module built using Go
(big Go, not TinyGo, although that would also work). This example
requires a few of these options to be configured - see the new
basic_test.go file for details. Specifically it requires:

- WASI enabled
- architecture set to wasm64 (because big Go compiles to wasm64)
- a custom module config with the `_initialize` function added to its
  list of start functions

The Go guest module is generated using [this branch] of `wit-bindgen-go`.
The custom branch is required because otherwise the generated Go code
doesn't compile, of limitations with Go's //go:wasmexport functionality.
There's a script (`gen-guest.sh`) that runs the `wit-bindgen-go`
command. Go compilation to Wasm is then done in the same
generate.go file as all the other examples.

[this branch]: https://github.com/bytecodealliance/go-modules/pull/367